refactor: ScrollToButon component
