+++ app/components/index.ts
@@ -15,3 +15,4 @@ export * from "./IconButton"
 export * from "./CarouselCard"
 export * from "./BoxShadow"
 export * from "./Carousel"
+export * from "./ScrollToButton"
